WiFiScanModify1_shj

+++ wifi/utils/src/wifi_common_util.cpp
@@ -225,7 +225,8 @@ unsigned int Ip2Number(const std::string& strIp)
 
     std::string ip(strIp + '.');
     while ((back = ip.find_first_of('.', back)) != (std::string::size_type)std::string::npos) {
-        number |= std::stol(ip.substr(front, back - front).c_str()) << (size -= sectionSize);
+        number |= static_cast<unsigned long>(std::stol(ip.substr(front, back - front).c_str()) <<
+            (size -= sectionSize));
         front = ++back;
     }
     return number;
